Fix adding travel and product to cart

+++ src/main/java/com/alaythiaproductions/hike_and_go/service/implementation/CartItemServiceImpl.java
@@ -31,9 +31,16 @@ public class CartItemServiceImpl implements CartItemService {
 
     @Override
     public CartItem updateCartItem(CartItem cartItem) {
-        BigDecimal subTotal = new BigDecimal(cartItem.getProduct().getOurPrice()).multiply(new BigDecimal(cartItem.getQty()));
-        subTotal = subTotal.setScale(2, BigDecimal.ROUND_HALF_UP);
+        BigDecimal subTotal = new BigDecimal(0);
 
+        if (cartItem.getProduct() != null) {
+            subTotal = new BigDecimal(cartItem.getProduct().getOurPrice()).multiply(new BigDecimal(cartItem.getQty()));
+            subTotal = subTotal.setScale(2, BigDecimal.ROUND_HALF_UP);
+        }
+        if (cartItem.getTravel() != null) {
+            subTotal = new BigDecimal(cartItem.getTravel().getOurPrice()).multiply(new BigDecimal(cartItem.getQty()));
+            subTotal = subTotal.setScale(2, BigDecimal.ROUND_HALF_UP);
+        }
         cartItem.setSubTotal(subTotal);
 
         return cartItemRepository.save(cartItem);
@@ -44,11 +51,13 @@ public class CartItemServiceImpl implements CartItemService {
         List<CartItem> cartItemList = findByShoppingCart(user.getShoppingCart());
 
         for (CartItem cartItem : cartItemList) {
-            if (product.getId() == cartItem.getProduct().getId()) {
-                cartItem.setQty(cartItem.getQty() + quantity);
-                cartItem.setSubTotal(new BigDecimal(product.getOurPrice()).multiply(new BigDecimal(quantity)));
-                cartItemRepository.save(cartItem);
-                return cartItem;
+            if (cartItem.getProduct() != null) {
+                if (product.getId() == cartItem.getProduct().getId()) {
+                    cartItem.setQty(cartItem.getQty() + quantity);
+                    cartItem.setSubTotal(new BigDecimal(product.getOurPrice()).multiply(new BigDecimal(quantity)));
+                    cartItemRepository.save(cartItem);
+                    return cartItem;
+                }
             }
         }
 
@@ -68,34 +77,36 @@ public class CartItemServiceImpl implements CartItemService {
         return cartItem;
     }
 
-//    @Override
-//    public CartItem addTravelToCartItem(Travel travel, User user, Integer quantity) {
-//        List<CartItem> cartItemList = findByShoppingCart(user.getShoppingCart());
-//
-//        for (CartItem cartItem : cartItemList) {
-//            if (travel.getId() == cartItem.getTravel().getId()) {
-//                cartItem.setQty(cartItem.getQty() + quantity);
-//                cartItem.setSubTotal(new BigDecimal(travel.getOurPrice()).multiply(new BigDecimal(quantity)));
-//                cartItemRepository.save(cartItem);
-//                return cartItem;
-//            }
-//        }
-//
-//        CartItem cartItem = new CartItem();
-//        cartItem.setShoppingCart(user.getShoppingCart());
-//        cartItem.setTravel(travel);
-//
-//        cartItem.setQty(quantity);
-//        cartItem.setSubTotal(new BigDecimal(travel.getOurPrice()).multiply(new BigDecimal(quantity)));
-//        cartItemRepository.save(cartItem);
-//
-//        TravelToCartItem travelToCartItem = new TravelToCartItem();
-//        travelToCartItem.setTravel(travel);
-//        travelToCartItem.setCartItem(cartItem);
-//        travelToCartItemRepository.save(travelToCartItem);
-//
-//        return cartItem;
-//    }
+    @Override
+    public CartItem addTravelToCartItem(Travel travel, User user, Integer quantity) {
+        List<CartItem> cartItemList = findByShoppingCart(user.getShoppingCart());
+
+        for (CartItem cartItem : cartItemList) {
+            if (cartItem.getTravel() != null) {
+                if (travel.getId() == cartItem.getTravel().getId()) {
+                    cartItem.setQty(cartItem.getQty() + quantity);
+                    cartItem.setSubTotal(new BigDecimal(travel.getOurPrice()).multiply(new BigDecimal(quantity)));
+                    cartItemRepository.save(cartItem);
+                    return cartItem;
+                }
+            }
+        }
+
+        CartItem cartItem = new CartItem();
+        cartItem.setShoppingCart(user.getShoppingCart());
+        cartItem.setTravel(travel);
+
+        cartItem.setQty(quantity);
+        cartItem.setSubTotal(new BigDecimal(travel.getOurPrice()).multiply(new BigDecimal(quantity)));
+        cartItemRepository.save(cartItem);
+
+        TravelToCartItem travelToCartItem = new TravelToCartItem();
+        travelToCartItem.setTravel(travel);
+        travelToCartItem.setCartItem(cartItem);
+        travelToCartItemRepository.save(travelToCartItem);
+
+        return cartItem;
+    }
 
     @Override
     public CartItem findById(Long cartItemId) {
@@ -104,7 +115,12 @@ public class CartItemServiceImpl implements CartItemService {
 
     @Override
     public void removeCartItem(CartItem cartItem) {
-        productToCartItemRepository.deleteByCartItem(cartItem);
+        if (cartItem.getProduct() != null) {
+            productToCartItemRepository.deleteByCartItem(cartItem);
+        }
+        if (cartItem.getProduct() != null) {
+            travelToCartItemRepository.deleteByCartItem(cartItem);
+        }
         cartItemRepository.delete(cartItem);
     }
